+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Modificação no Middleware adicionado class admin 
 - Modificação no codigo do arquivo controller (TipoUsuario) 
 # [1.0.0] - 2025-02-14
+- Adicionado modal para pagina de produto